* master: (54 commits)
  [ChannelSelection] * correct typo from the last commit
  Ellipsis (#2892)
  pot update
  Prepare github actions for the new 7.3 branch
  ChannelSelection fixes and enhancements (#2894)
  [ChoiceList.py] Correct location of bullet icon
  [HarddiskSetup] Change "Initialize" to "Format" A storage device is initialised automatically when you record, go to flash online etc without any user interevention. The context of the menu/ question is that it is asking if we want to format the device, not if we want to initialise it. (initail setup for use). Most E2 users are from a Windows background and "Format" in this context reflects the function in a more user friendly term.
  prepare new Channelselection
  translation update
  [ChannelSelection.py] Clean up code (#2893)
  [Harddisk] * remove sfdisk usage because it's no longer working
  [HarddiskSetup] * small cleanup * fix running in Wizard
  [Components/Network]
  aml lcd display set display brightness to 0 in standby
  [StartEnigma.py] Correct last commit
  Fix ParentalControl import issue (#2888)
  translation update
  [PluginBrowser] * remove callbacks on close
  backup exclude /run/chrony/chronyd.sock
  [Network] (#2885)
  ...

+++ lib/python/Plugins/SystemPlugins/SoftwareManager/ImageBackup.py
@@ -170,6 +170,7 @@ class ImageBackup(Screen):
 				self.IMGREVISION = BoxInfo.getItem("imgrevision")
 				self.IMGVERSION = BoxInfo.getItem("imgversion")
 				self.KERNEL = BoxInfo.getItem("kernel")
+				self.HasKexecMultiboot = BoxInfo.getItem("HasKexecMultiboot")
 
 				if BoxInfo.getItem("canRecovery"):
 					self.EMMCIMG = BoxInfo.getItem("canRecovery")[0]
@@ -213,8 +214,12 @@ class ImageBackup(Screen):
 				print("[Image Backup] USB RECOVERY = >%s< " % self.RECOVERY)
 				print("[Image Backup] DESTINATION = >%s< " % self.DIRECTORY)
 				print("[Image Backup] SLOT = >%s< " % self.SLOT)
+				print("[Image Backup] HasKexecMultiboot = >%s< " % self.HasKexecMultiboot)
+				print("[Image Backup] canMultiBoot = >%s< " %MultiBoot.canMultiBoot())
 
 				isNotCurrent = MultiBoot.getCurrentSlotCode() != answer[1]
+
+				print("[Image Backup] isNotCurrent = >%s< " %isNotCurrent)
 
 				if self.RECOVERY and not isNotCurrent:
 					print("[Image Backup] IMAGEDISTRO = >%s<" % self.DISTRO)
@@ -232,7 +237,7 @@ class ImageBackup(Screen):
 				self.DATE = strftime("%Y%m%d_%H%M", localtime(self.START))
 				self.MKFS_UBI = "/usr/sbin/mkfs.ubifs"
 				self.MKFS_TAR = "/bin/tar"
-				self.BZIP2 = "/usr/bin/bzip2"
+				self.BZIP2 = "/usr/bin/pbzip2"
 				self.MKFS_JFFS2 = "/usr/sbin/mkfs.jffs2"
 				self.UBINIZE = "/usr/sbin/ubinize"
 				self.NANDDUMP = "/usr/sbin/nanddump"
@@ -351,7 +356,7 @@ class ImageBackup(Screen):
 					cmd1 = "%s -r %s -o %s/root.ubi %s" % (self.MKFS_UBI, self.backuproot, self.WORKDIR, self.MKUBIFS_ARGS)
 					cmd2 = "%s -o %s/root.ubifs %s %s/ubinize.cfg" % (self.UBINIZE, self.WORKDIR, self.UBINIZE_ARGS, self.WORKDIR)
 				elif not self.RECOVERY:
-					cmd1 = "%s -cf %s/rootfs.tar -C %s --exclude ./var/nmbd --exclude ./.resizerootfs --exclude ./.resize-rootfs --exclude ./.resize-linuxrootfs --exclude ./.resize-userdata --exclude ./var/lib/samba/private/msg.sock --exclude ./var/lib/samba/msg.sock/* --exclude ./run/avahi-daemon/socket ." % (self.MKFS_TAR, self.WORKDIR, self.backuproot)
+					cmd1 = "%s -cf %s/rootfs.tar -C %s --exclude ./var/nmbd --exclude ./.resizerootfs --exclude ./.resize-rootfs --exclude ./.resize-linuxrootfs --exclude ./.resize-userdata --exclude ./var/lib/samba/private/msg.sock --exclude ./var/lib/samba/msg.sock/* --exclude ./run/avahi-daemon/socket --exclude ./run/chrony/chronyd.sock ." % (self.MKFS_TAR, self.WORKDIR, self.backuproot)
 					cmd2 = "sync"
 					cmd3 = "%s %s/rootfs.tar" % (self.BZIP2, self.WORKDIR)
 
@@ -422,7 +427,10 @@ class ImageBackup(Screen):
 
 				cmdlist.append(self.makeEchoCreate("kerneldump"))
 				if MultiBoot.canMultiBoot() or self.MTDKERNEL.startswith("mmcblk0") or self.MACHINEBUILD in ("h8", "hzero"):
-					cmdlist.append("dd if=/dev/%s of=%s/%s" % (self.MTDKERNEL, self.WORKDIR, self.KERNELBIN))
+					if BoxInfo.getItem("HasKexecMultiboot"):
+						cmdlist.append("cp /%s %s/%s" % (self.MTDKERNEL, self.WORKDIR, self.KERNELBIN))
+					else:
+						cmdlist.append("dd if=/dev/%s of=%s/%s" % (self.MTDKERNEL, self.WORKDIR, self.KERNELBIN))
 				else:
 					cmdlist.append("nanddump -a -f %s/vmlinux.gz /dev/%s" % (self.WORKDIR, self.MTDKERNEL))
